moved PageValidator
